refactor(SDK): replace OMSDK by the prebid's version
