--- PrebidMobile/Rendering/PrebidMobileRendering/Frameworks/OMSDK_Openx.framework/OMSDK_Openx
@@ -1,1 +1,0 @@
-Versions/Current/OMSDK_Openx
