--- PrebidMobile/Rendering/PrebidMobileRendering/Frameworks/OMSDK_Openx.framework/Versions/Current
@@ -1,1 +1,0 @@
-A
