+++ PrebidMobile/Rendering/PrebidMobileRendering/Frameworks/OMSDK_Prebidorg.framework/Modules/module.modulemap
@@ -1,0 +1,4 @@
+module OMSDK_Prebidorg {
+    header "Headers/OMIDImports.h"
+    export *
+}
