+++ PrebidMobile/Rendering/PrebidMobileRendering/PrebidMobileRendering/3dPartyWrappers/OpenMeasurement/PBMOpenMeasurementFriendlyObstructionTypeBridge.h
@@ -8,7 +8,7 @@
 #import <Foundation/Foundation.h>
 #import "PBMOpenMeasurementFriendlyObstructionPurpose.h"
 
-@import OMSDK_Openx;
+@import OMSDK_Prebidorg;
 
 NS_ASSUME_NONNULL_BEGIN
